Added the version check. Fix some warning. Fix a double free after a snapshot was restored.

+++ Ponce/src/actions.cpp
@@ -240,7 +240,7 @@ struct ah_taint_memory_t : public action_handler_t
 				{
 					auto selection_starts = action_update_ctx_t->cur_sel.from.at->toea();
 					auto selection_ends = action_update_ctx_t->cur_sel.to.at->toea();
-					int diff = selection_ends - selection_starts;
+					int diff = (int)(selection_ends - selection_starts);
 					if (diff >= 0)
 						return AST_ENABLE;
 				}
@@ -340,7 +340,7 @@ struct ah_symbolize_memory_t : public action_handler_t
 				{
 					auto selection_starts = action_update_ctx_t->cur_sel.from.at->toea();
 					auto selection_ends = action_update_ctx_t->cur_sel.to.at->toea();
-					int diff = selection_ends - selection_starts;
+					int diff = (int)(selection_ends - selection_starts);
 					if (diff >= 0)
 						return AST_ENABLE;
 				}
